refactor: introduce shared inbound-turn submit primitive
Extract shared request parsing/validation from both POST /v1/runs and
POST /v1/messages into send-validation.ts. Both handlers now delegate
to parseSendRequest() for body parsing, field validation, attachment
resolution, and conversation lookup.

No external behavior changes.
